HummingBird: drastic changes to how URL query params are handled.
Originally I was encoding the PostgREST query string and passing that to the URL params.

Now I am just encoding an object, then when I refresh the page I decode the object and convert it to a query string at that point.

This is a better approach because it's more versatile: it can be used for the other query params.

In particular I have started using "v" to encode all the view params. This will be one big object that includes modifications to the columns and view specific information (e.g. the pivot column in the Kanban view)