HummingBird: adding basic support for history.
This will need a lot more work. I think ultimately this is going to be an external service that users can call from outside. If there is only a record history from Pollygot there there will be missing info.
We should allow users to send the record changes to Pollygot, with an author "API" or "Username" etc.